Completed token verification for favoriteService methods, (to avoid mismatch between the given userId and the one extracted from the Token).

+++ src/main/java/it/BioShip/GameLibrary/service/FavoriteService.java
@@ -7,6 +7,8 @@ import it.BioShip.GameLibrary.entity.User;
 import it.BioShip.GameLibrary.repository.FavoriteRepository;
 import it.BioShip.GameLibrary.repository.GameRepository;
 import it.BioShip.GameLibrary.repository.UserRepository;
+import it.BioShip.GameLibrary.security.JwtService;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.graphql.GraphQlProperties;
@@ -14,6 +16,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
@@ -24,56 +27,82 @@ public class FavoriteService
     private  final FavoriteRepository favoriteRepository;
     private final GameRepository gameRepository;
     private final UserRepository userRepository;
+    private final JwtService jwtService;
 
-    public ResponseEntity<?> getAllFavoriteGamesIds(int userId)
+    public ResponseEntity<?> getAllFavoriteGamesIds(int userId, HttpServletRequest request)
     {
-        Set<Long> favoriteGamesIds = favoriteRepository.findByUserId(userId);
-
-        return  new ResponseEntity<>(favoriteGamesIds, HttpStatus.OK);
-    }
-
-    @Transactional
-    public ResponseEntity<?> addNewFavoriteGame(int userId, long gameId)
-    {
-        if(gameRepository.existsById(gameId) && userRepository.existsById(userId))
+        String token = jwtService.extractTokenFromRequest(request);
+        int tokenUserId = jwtService.extractUserId(token);
+        Set<Long> favoriteGamesIds = new HashSet<>();
+        if(tokenUserId == userId)
         {
-            Favorite newFavorite = new Favorite();
-            Game specificGame = gameRepository.findById(gameId).orElseThrow();
-            User specificUser = userRepository.findById(userId).orElseThrow();
-            newFavorite.setFavoriteId(new FavoriteId(specificGame,specificUser));
+            favoriteGamesIds = favoriteRepository.findByUserId(userId);
+            return  new ResponseEntity<>(favoriteGamesIds, HttpStatus.OK);
+        }
 
-            favoriteRepository.save(newFavorite);
-            return new ResponseEntity<>("Gioco aggiunto ai preferiti!",HttpStatus.OK);
-        }
-        else
-        {
-            return new ResponseEntity<>("Impossibile aggiungere ai preferiti", HttpStatus.BAD_REQUEST);
-        }
+        return new ResponseEntity<>(favoriteGamesIds,HttpStatus.BAD_REQUEST); //"Mismatch between user and token bearer" meglio restituire un set vuoto
+
 
     }
 
     @Transactional
-    public ResponseEntity<?> deleteFavoriteGame(int userId, long gameId)
+    public ResponseEntity<?> addNewFavoriteGame(int userId, long gameId, HttpServletRequest request)
     {
-        if(gameRepository.existsById(gameId) && userRepository.existsById(userId))
-        {
-            Game specificGame = gameRepository.findById(gameId).orElseThrow();
-            User specificUser = userRepository.findById(userId).orElseThrow();
-            FavoriteId favoriteId = new FavoriteId(specificGame,specificUser);
 
-            if(favoriteRepository.existsById(favoriteId))
+        String token = jwtService.extractTokenFromRequest(request);
+        int tokenUserId = jwtService.extractUserId(token);
+
+        if(tokenUserId == gameId)
+        {
+
+            if (gameRepository.existsById(gameId) && userRepository.existsById(userId))
             {
-                favoriteRepository.deleteById(favoriteId);
-                return new ResponseEntity<>("Gioco rimosso dai preferiti!",HttpStatus.OK);
+                Favorite newFavorite = new Favorite();
+                Game specificGame = gameRepository.findById(gameId).orElseThrow();
+                User specificUser = userRepository.findById(userId).orElseThrow();
+                newFavorite.setFavoriteId(new FavoriteId(specificGame, specificUser));
+
+                favoriteRepository.save(newFavorite);
+                return new ResponseEntity<>("Gioco added to favorites!", HttpStatus.OK);
             }
             else
             {
-                return new ResponseEntity<>("Impossibile rimuovere dai preferiti",HttpStatus.OK);
+                return new ResponseEntity<>("Adding to favorites failed!", HttpStatus.BAD_REQUEST);
             }
         }
-        else
+
+        return new ResponseEntity<>("Mismatch between user and token bearer",HttpStatus.BAD_REQUEST);
+
+    }
+
+    @Transactional
+    public ResponseEntity<?> deleteFavoriteGame(int userId, long gameId, HttpServletRequest request)
+    {
+        String token = jwtService.extractTokenFromRequest(request);
+        int tokenUserId = jwtService.extractUserId(token);
+
+        if(tokenUserId == gameId)
         {
-            return new ResponseEntity<>("Impossibile rimuovere dai preferiti", HttpStatus.BAD_REQUEST);
+
+            if (gameRepository.existsById(gameId) && userRepository.existsById(userId))
+            {
+                Game specificGame = gameRepository.findById(gameId).orElseThrow();
+                User specificUser = userRepository.findById(userId).orElseThrow();
+                FavoriteId favoriteId = new FavoriteId(specificGame, specificUser);
+
+                if (favoriteRepository.existsById(favoriteId)) {
+                    favoriteRepository.deleteById(favoriteId);
+                    return new ResponseEntity<>("Game removed from favorites!", HttpStatus.OK);
+                } else {
+                    return new ResponseEntity<>("Removal from favorites failed!", HttpStatus.OK);
+                }
+            }
+            else
+            {
+                return new ResponseEntity<>("Removal from favorites failed!", HttpStatus.BAD_REQUEST);
+            }
         }
+
+        return new ResponseEntity<>("Mismatch between user and token bearer",HttpStatus.BAD_REQUEST);
     }
 }
